refactor(store): extract macros domain into a zustand slice (#2114)
Move the Macros domain — the stored-macro library plus recording (#1674)
and playback (#1675) state and their actions — out of the appStore.ts
monolith into a dedicated StateCreator slice at
src/store/slices/macrosSlice.ts, spread into the root create() exactly as
the SSH tunnel (#2077) and embedded-servers (#2113) slices are.

Behavior-preserving pure structural move: the slice's set/get stay typed
against the full AppState, so the public store shape and every action
signature/behavior are unchanged. The module-local generateMacroId helper
and the activeMacroPlayback handle move with the domain; MacroPlaybackState
and PlayMacroOptions move to the slice and are re-exported from appStore to
keep its export surface identical. collectLiveTabs is now exported so the
slice's playMacro can reuse it alongside getActiveTab.

No macros action creates a tab, so nothing is left behind in the root store
(unlike tunnels' openTunnelEditorTab). runMacroPlayback/getTerminalInputInjector
stay imported in appStore.ts because the workflow runner also uses them.

Verified: full frontend suite (497 files / 4665 tests) green, tsc + eslint +
prettier clean.
